+++ src/canister_types/src/models/notification.rs
@@ -7,7 +7,7 @@ use candid::{Decode, Encode};
 use crate::impl_storable_for;
 
 use super::{
-    attendee::InviteAttendeeResponse,
+    attendee::{InviteAttendeeResponse, JoinedAttendeeResponse},
     friend_request::FriendRequestResponse,
     member::{InviteMemberResponse, JoinedMemberResponse},
     user_notifications::UserNotificationData,
@@ -93,10 +93,12 @@ pub enum GroupNotificationType {
     JoinGroupOwnerRequestAccept(InviteMemberResponse),
     JoinGroupOwnerRequestDecline(InviteMemberResponse),
 
-    MemberRoleAssign(JoinedMemberResponse),
-    UserRemoveInvite(InviteMemberResponse),
-    UserJoinGroup(u64),
+    RoleAssignByOwner(JoinedMemberResponse),
+    RemoveInviteByOwner(InviteMemberResponse),
+    RemoveMemberByOwner(JoinedMemberResponse),
+
     UserLeaveGroup(u64),
+    UserJoinGroup(u64),
     GroupReminder(u64),
 }
 
@@ -112,8 +114,12 @@ pub enum EventNotificationType {
     JoinEventOwnerRequestAccept(InviteAttendeeResponse),
     JoinEventOwnerRequestDecline(InviteAttendeeResponse),
 
-    UserJoinEvent(u64),
-    UserLeaveEvent(u64),
+    RoleAssignByOwner(JoinedAttendeeResponse),
+    RemoveInviteByOwner(InviteAttendeeResponse),
+    RemoveAttendeeByOwner(JoinedAttendeeResponse),
+
+    UserJoinEvent((u64, u64)),
+    UserLeaveEvent((u64, u64)),
     EventReminder(u64),
 }